+++ src/index.ts
@@ -1,3 +1,10 @@
 export * from "./client/client";
-export { HandlerFunc, PublishEvent, SubscriptionEvent } from "./client/event";
+export {
+    HandlerFunc,
+    PublishEvent,
+    SubscriptionEvent,
+    BaseEvent,
+    EventData,
+    GdprEventData,
+} from "./client/event";
 export { ClientConfig } from "./client/config";